backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-channels==0.7.0
django-mptt==0.16.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
redis==5.2.1
daphne==4.1.2
channels-redis==4.2.1
Pillow==11.1.0 